Add multiple interface support
