+++ Interfaces.h
@@ -1,0 +1,7 @@
+#define SX127X 0x00
+#define SX1276 0x01
+#define SX1278 0x02
+#define SX126X 0x10
+#define SX1262 0x11
+#define SX128X 0x20
+#define SX1280 0x21
